+++ editor/src/main/java/org/nmox/studio/editor/typing/JsDeletedTextInterceptor.java
@@ -100,7 +100,8 @@ public class JsDeletedTextInterceptor implements DeletedTextInterceptor {
         @MimeRegistration(mimeType = "text/x-crystal", service = DeletedTextInterceptor.Factory.class),
         @MimeRegistration(mimeType = "text/css", service = DeletedTextInterceptor.Factory.class),
         @MimeRegistration(mimeType = "text/x-scss", service = DeletedTextInterceptor.Factory.class),
-        @MimeRegistration(mimeType = "text/x-less", service = DeletedTextInterceptor.Factory.class)
+        @MimeRegistration(mimeType = "text/x-less", service = DeletedTextInterceptor.Factory.class),
+        @MimeRegistration(mimeType = "text/x-svelte", service = DeletedTextInterceptor.Factory.class)
     })
     public static class Factory implements DeletedTextInterceptor.Factory {
         @Override